WORKING BASE FLAVOR AND SCOOP NUMBER!!!
I got the code to work so the user is prompted to input their choice of base flavor and scoop amount. If they enter a non-number character or a number other than 1, 2, or 3, they are told that the input is invalid and prompted again until they input an accepted value.